refactor: move functions to actual `bal_addresses` module
